+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+Drop symfony/polyfill-uuid in favour of a standalone implementation (#1346)
+
 ## 3.7.0 (2022-07-18)
 
 - Fix `Scope::getTransaction()` so that it returns also unsampled transactions (#1334)